Fixed EditView CustomTextField
For whatever reason at some point it just lost the ability to visually update anything inserted. Like, it accepted the input, but immediately reverted to the state the Habit was created in.

I have no clue on neither what caused it nor how I fixed it.

Still no autosave functionality.