EditView Mark Completed Button functionality 2
Checkboxes' visuals now change based on the actual db status of a given Habit. And this status is updated in accordance to progress and goal.

The issue is that the logic for updating the status works exactly the opposite way to what I expect and it only works for Binary Habits.